Connect DOM through scripts
